requirements.txt
tensorflow
scikit-learn
numpy
pandas
python-binance
TA-Lib